Complete architectural cleanup
- Remove vestigial Agent Executor functions (initialize_thudbot, get_thud_agent)
- Move find_hint_node function to its own file for proper modular structure
- Update imports and tests to reflect new RAG-only architecture
- System now fully uses initialize_rag_only() and get_direct_hint() approach
