+++ src/__init__.py
@@ -1,7 +1,7 @@
 # src/__init__.py
 
 # Import the Thudbot agent components
-from .agent import get_thud_agent, initialize_thudbot
+from .agent import get_direct_hint, initialize_rag_only
 from .api import app
 
-__all__ = ['get_thud_agent', 'initialize_thudbot', 'app']
+__all__ = ['get_direct_hint', 'initialize_rag_only', 'app']
